+++ Publish-GitPubJekyll.ps1
@@ -42,6 +42,11 @@ function Publish-GitPubJekyll {
     [Parameter(ValueFromPipelineByPropertyName)]
     [string]
     $PostLayout,
+    # The source URL.  If provided, this will be included in front matter.
+    [Parameter(ValueFromPipelineByPropertyName)]
+    [Alias('HTML_url')]
+    [string]
+    $SourceUrl,
     # The output path.  If not provided, will output to _posts in the current directory.
     [string]
     $OutputPath    
@@ -72,6 +77,9 @@ function Publish-GitPubJekyll {
         }
         if ($PostAuthor) {
             $frontMatter['author'] = $PostAuthor
+        }
+        if ($SourceUrl) {
+            $frontMatter['sourceURL'] = $SourceUrl
         }
         if ($PostTag) {
             if ($PostTag.Length -eq 1) {